Fixed 3 block error and much more. Added TODO
TO DO LIST:
1. Modify end game condition (!endgame)
2. Modify Input decoder
a. At the moment, I have it choosing a coordinate. If Dave’s website
were up at the moment, I would copy and paste the UP/DOWN/LEFT/RIGHT
chart he had.
b. The same movement algorithm can be applied to Blocks and Developers
i. Instead of x, y coordinates, make it that pressing ‘8’ moves the
piece up etc (maybe from the middle of the board)
3. Differentiate what 1 block tile is placed. At the moment, I am
treating them both the same
a. Prompt for what Tile the user wants to select
b. It was late last night and couldn’t figure out how to pass the
parameter is instanceOf Village or is instanceOf Irrigation
4. Scoring needs to be taken care of.
a. Case 4 in the switch statement needs some love. Building Palace
method
b. When placing any Tile, the scoring method needs to be checked (is
there a surrounded irrigation tile?) Fun Algorithms here!
5. For Case 1 and 2, I SHOULD be able to print out only the developers
in the inventory or the players on the board. That was the intention at
least. Right now it displays them all.
My position class is funky. In Board, I made a method that is
isThereDeveloper() and it has a funky thing to it. I’ll be honest I’m
not sure why I did
public static boolean isThereDeveloper(Space s) { but it worked
6.
OPTIONAL:
1. Deduct Action point when block is placed. At the moment, no action
point is taken.
2. Separate some methods into the mediator class (some might get mad at
me for saying this is optional?
3. Find a way to display the land type and developers at the same time.
(Maybe even height)